chore: rename provider files
