+++ chatlas/_provider_databricks.py
@@ -4,12 +4,12 @@ from typing import TYPE_CHECKING, Optional
 
 from ._chat import Chat
 from ._logging import log_model_default
-from ._openai import OpenAIProvider
+from ._provider_openai import OpenAIProvider
 
 if TYPE_CHECKING:
     from databricks.sdk import WorkspaceClient
 
-    from ._openai import ChatCompletion
+    from ._provider_openai import ChatCompletion
     from .types.openai import SubmitInputArgs
